update compute type

+++ src/faster_whisper_app/config.py
@@ -17,7 +17,7 @@ class Config:
     # Model settings (optimized for speed)
     model_size: str = "small"  # Changed from "base" for 2x speed improvement
     device: str = "cpu"
-    compute_type: str = "float16"  # Changed from "int8" for better speed/quality balance
+    compute_type: str = "int8"  # Optimized for CPU - much faster than float16 on CPU
     
     # Audio settings
     sample_rate: int = 16000
@@ -40,7 +40,7 @@ def load_config() -> Config:
     return Config(
         model_size=os.getenv("FASTER_WHISPER_MODEL_SIZE", "small"),  # Default to "small" for speed
         device=os.getenv("FASTER_WHISPER_DEVICE", "cpu"),
-        compute_type=os.getenv("FASTER_WHISPER_COMPUTE_TYPE", "float16"),  # Default to "float16"
+        compute_type=os.getenv("FASTER_WHISPER_COMPUTE_TYPE", "int8"),  # Default to "int8" for CPU
         sample_rate=int(os.getenv("AUDIO_SAMPLE_RATE", "16000")),
         channels=int(os.getenv("AUDIO_CHANNELS", "1")),
         audio_device_index=int(os.getenv("AUDIO_DEVICE_INDEX")) if os.getenv("AUDIO_DEVICE_INDEX") else None,
